refactor(db): 修改dish, takeaway_dish, takeaway_shop的外键为integer类型

+++ backend/base/src/models/takeaway_dish.rs
@@ -11,7 +11,7 @@ pub struct Model {
     pub dish_type: String,
     #[sea_orm(column_type = "Decimal(Some((10, 2)))")]
     pub price: Decimal,
-    pub takeaway_shop_id: i64,
+    pub takeaway_shop_id: i32,
     pub images: Json,
 }
